Back @State and @AppStorage with Compose MutableState, save in Saver
Saving @State and @AppStorage in the Saver allows us to support bindings to state in views that disappear
from the Compose tree - e.g. because they are on the back stack. Backing these with MutableState means that
bindings can "see" and recompose on changes to the values, because when the binding reads the value its
composition context will automatically get registered to recompose on change